+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 1.15.0 (2017-04-13)
 
   * Add '/health' endpoint to retrieve information about appliance health
+  * Display health status at top Admin Dashboard
   * Return jqXHR on API call errors for easier debugging
+  * Fix disk usage graph display. #26
 
 ## 1.14.0 (2017-02-10)